Added logger capability for errors

+++ external-libs/bson/binary.cc
@@ -163,7 +163,7 @@ Handle<Value> Binary::ReadInto(const Arguments &args) {
   
   size_t offset = args[1]->IsUint32() ? args[1]->Uint32Value() : 0;
   Binary *binary = ObjectWrap::Unwrap<Binary>(args.This());
-  if (binary->index - offset < 0) {
+  if(((int)binary->index - (int)offset) < 0) {
      return VException("offset argument out of bounds"); 
   }
